requirements.txt
requests
flask
scikit-learn
pandas
numpy
tweepy
pyyaml